+++ design_docs/history.md
@@ -50,5 +50,4 @@
 10/02/25
 
 - HexGrid work : Refactored Gen + Circle Algorithm 9-10
-- HexGrid Circle Algorithm 10-11  11-12 14-15 15-16
--
+- HexGrid Circle Algorithm 10-11  11-12 14-15 15-16 16-17